moved code from resources to external to fit with existing file structure
